frontend: rename qt and android files
So that these files exist next to each other.
